remoteproc/pru: Deprecate usage of pruss_intc_trigger()
The only user of the pruss_intc_trigger() API is the PRU remoteproc
driver itself. Use the API irq_set_irqchip_state() provided by the
kernel IRQ subsystem directly instead. This is in preparation for
dropping the pruss_intc_trigger() API and the corresponding header
fil, and is the recommended way going forward for any new PRU client
drivers as well.

Signed-off-by: Suman Anna <s-anna@ti.com>